+++ SADXModManager/Forms/ConfigEditDialog.resx
