+++ SADXModManager/Forms/NewModDialog.resx
